Move new tests under e2e_new
